+++ tests/phpunit/tests/classes/hook/action/group/activity/update.php
@@ -125,7 +125,7 @@ class WordPoints_BP_Hook_Action_Group_Activity_Update_Test
 			'test_event'
 			, array( 'a', $activity )
 			, array(
-				'arg_index' => array( 'bp_group_activity_update' => 1 ),
+				'arg_index'    => array( 'bp_group_activity_update' => 1 ),
 				'requirements' => array( 0 => 'a' ),
 			)
 		);
@@ -150,7 +150,7 @@ class WordPoints_BP_Hook_Action_Group_Activity_Update_Test
 			'test_event'
 			, array( 'a', $activity )
 			, array(
-				'arg_index' => array( 'bp_group_activity_update' => 1 ),
+				'arg_index'    => array( 'bp_group_activity_update' => 1 ),
 				'requirements' => array( 0 => 'b' ),
 			)
 		);